Ensure unit and doc tests compile and run.
Previously, running `$ cargo test --all` would fail with doctest
compilation issues as well as unit test failures. This change ensures
that the doctests compile and the unit test is updated.

The unit test in question (`test_table_insertion_with_list`) was
failing due to the generated SQL having `CAST` statements in it - the
`assert_eq` has been updated.

This change also adds a `test` make target that will run the rust
tests and ensures that this is run as a part of the PR GHA.